backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
youtube-transcript-api==0.6.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
together==0.2.11 